+++ test/models/custom_field_test.rb
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 # = Informations
 #
 # == License
@@ -8,16 +9,16 @@
 # Copyright (C) 2012-2014 Brice Texier, David Joulin
 #
 # This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
+# it under the terms of the GNU Affero General Public License as published by
 # the Free Software Foundation, either version 3 of the License, or
 # any later version.
 #
 # This program is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
+# GNU Affero General Public License for more details.
 #
-# You should have received a copy of the GNU General Public License
+# You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see http://www.gnu.org/licenses.
 #
 # == Table: custom_fields
@@ -47,19 +48,24 @@ require 'test_helper'
 class CustomFieldTest < ActiveSupport::TestCase
 
   for model in Ekylibre::Record::Base.descendants
-    test "new custom field on #{model.name}" do
-      for nature in CustomField.nature.values
-        field = CustomField.create!(:name => "A #{nature} info", :nature => nature, :customized_type => model.name)
-        assert model.connection.columns(model.table_name).detect{|c| c.name.to_s == field.column_name}
+    test "manage custom field on #{model.name.underscore}" do
+      if CustomField.customized_type.values.include?(model.name)
+        assert_raise ActiveRecord::RecordInvalid, "Cannot add custom field on not customizable models like #{model.name}" do
+          CustomField.create!(name: "たてがみ", nature: :text, customized_type: model.name)
+        end
+      else
+        for nature in CustomField.nature.values
+          field = CustomField.create!(name: "#{nature.capitalize} info", nature: nature, customized_type: model.name)
+          assert model.connection.columns(model.table_name).detect{|c| c.name.to_s == field.column_name}
 
-        field.name = "A #{nature} info"
-        field.save
-        assert model.connection.columns(model.table_name).detect{|c| c.name.to_s == field.column_name}
+          field.name = "#{nature.capitalize} インフォ"
+          field.save
+          assert model.connection.columns(model.table_name).detect{|c| c.name.to_s == field.column_name}
 
-
-        column_name = field.column_name
-        field.destroy
-        assert !model.connection.columns(model.table_name).detect{|c| c.name.to_s == column_name}
+          column_name = field.column_name
+          field.destroy
+          assert !model.connection.columns(model.table_name).detect{|c| c.name.to_s == column_name}
+        end
       end
     end
   end
